// src/app/shared/models/Clientes.ts
export interface IClientes {
  GFCLINUM: string;
  GFCLINOM: string;
  GFCLINO1: string;
  GFCLINO2: string;
  GFCLIAP1: string;
  GFCLIAP2: string;
  GFCLITEL: string;
  GFCLIDIR: string;
  GRUPO: string;
  GFCLICUR: string;
  GFCLIFNA: Date;
  GFCLIINE: string;
  PAIS: string;
  TAPAICOD: string;
  TADEPNOM: string;
  TADEPCOD: string;
  TAMUNNOM: string;
  TAMUNCOD: string;
  TABARNOM: string;
  TABARCOD: string;
  SUCURSALGRUP: string;
  CICLO: string;
  OFICIAL: string;
}